docs: add an evidence map, contributing guide and review scaffolding
docs/EVIDENCE.md maps each assessed practice to the files that implement it and
a command that checks it rather than taking our word for it — including a
closing section on what is deliberately not done: no staging environment, a
single host, dev-mode Vault, no paging destination.

CONTRIBUTING.md separates the rules that are enforced by a gate from the ones
that are merely suggested. CODEOWNERS routes review by blast radius: the money
path, anything holding key material, and the pipeline itself.

.githooks/pre-commit is deliberately not a mirror of CI — a hook that takes 40
seconds gets bypassed, and a bypassed hook protects nothing. It checks staged
files for leaked credentials and formatting only.

Branch protection is a repository setting, not a file, so
scripts/setup-branch-protection.sh makes the intended configuration explicit and
re-appliable instead of invisible.

The DR runbook gains the two drills it was missing: restore from backup, and
release rollback.

+++ services/risk-ai-service/app/main.py
@@ -9,11 +9,11 @@ from uuid import uuid4
 from fastapi import FastAPI, HTTPException
 from pydantic import BaseModel, Field
 
+from app.federated import fedavg_demo
+from app.metrics import setup_metrics
 from app.model import RiskModel, amount_z_score
 from app.rules import Decision, decide, rules_engine
 from app.shield import AiShield, ServiceHealthSample
-from app.federated import fedavg_demo
-from app.metrics import setup_metrics
 
 ARTIFACT_DIR = Path(os.environ.get("FINIX_MODEL_DIR", Path(__file__).parent / "model_artifacts"))
 
@@ -95,7 +95,7 @@ def score_transaction(body: ScoreRequest) -> ScoreResponse:
         hour=hour,
     )
     # Blend: rules dominate for explainability; model nudges.
-    blended = int(round(0.65 * rules_score + 0.35 * model_score))
+    blended = round(0.65 * rules_score + 0.35 * model_score)
     blended = max(0, min(100, blended))
     decision = decide(blended)
     case_id = None
